Changed double to decimal for entire application. Added coin count for exact change.

+++ VendingMachineKata/Models/Item.cs
@@ -10,9 +10,9 @@ namespace VendingMachineKata.Models
     {
         public int ID { get; set; }
         public string Name { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
-        public Item(int id, string name, double price)
+        public Item(int id, string name, decimal price)
         {
             ID = id;
             Name = name;
